+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename datasources to get rid of the olly-op part.
+
 ## [0.11.0] - 2025-01-10
 
 ### Added